+++ packages/core/src/index.ts
@@ -102,13 +102,13 @@ export * from './common/view/text';
 export * from './common/view/view';
 
 export * from './common/number/util';
+export * from './common/point-nd/util';
 export * from './common/compat';
 export * from './common/dom-util';
 export * from './common/list-util';
 export * from './common/micro-parsers';
 export * from './common/params';
 export * from './common/picker-util';
-export * from './common/point-nd-util';
 export * from './common/primitive';
 export * from './common/tp-error';
 export * from './common/ui';
